+++ CHANGELOG.md
@@ -18,4 +18,17 @@
 ### Known Issues
 -->
 
+## 0.6.2
+
+### Added
+- Added an option to the kit's right-click context menu to set the preferred kit
+- Added an option to the ability's right click context menu to swap between melee/ranged usage on weapons with a distance of melee or ranged.
+- Added an embed for kits used for viewing on the actor sheet, viewing in play mode on the kit sheet, and posting to chat. (#239)
+
+### Changed
+- Updated the view of the features tab to match the style of the abilities tab
+- The kit section will only be visible if the actor's class allows kits or there's kits on the actor.
+- Made the kit swap dialog more clear on why you need to swap kits.
+
+
 ## 0.6 Initial Release